Move GlobalConfig into API
